requirements.txt
gradio
torch
transformers
pillow